requirements.txt
# Truth Social Trump Monitor Dependencies

# Web 框架
fastapi>=0.109.0
uvicorn>=0.27.0

# HTTP 客户端
httpx>=0.25.0
requests>=2.31.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# 定时任务
apscheduler>=3.10.0

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 时间处理
pytz>=2024.1

# 重试机制
tenacity>=8.2.0

# 日志
loguru>=0.7.0

# LLM 预留（可选）
openai>=1.0.0
